almost finiished

+++ controllers/AlbumController.php
@@ -21,71 +21,77 @@ class AlbumController extends Controller
         if (isset ( $_SESSION ['loggedIn'] ) && $_SESSION ['loggedIn'] == true) {
             $photoTagModel = new PhotoTagModel();
             $photoAlbumModel = new PhotoAlbumModel();
+            $userModel = new UserModel();
             $albumModel = new AlbumModel();
             $tagModel = new TagModel();
-            $photos = $photoAlbumModel->readAllPhotosByAlbumId($id);
-            $album = $albumModel->readById($id);
 
-            $tagIds = isset ( $_GET ['tags'] ) ? $_GET ['tags'] : "";
+            if ($albumModel->readIsAlbumFromUser($id,$userModel->readIdByUsername($_SESSION['userName']))) {
+                $photos = $photoAlbumModel->readAllPhotosByAlbumId($id);
+                $album = $albumModel->readById($id);
 
-            $tagArray = explode('-',$tagIds);
-            $tagArray =  array_unique($tagArray);
-            $selectedTags = array();
-            foreach ($tagArray as $tagId) {
-                if ($tagId != "") {
-                    if($tagModel->readById($tagId)) {
-                        array_push($selectedTags, $tagModel->readById($tagId));
-                    }
-                }
-            }
-            $selectedPhotos = array();
-            if(count($selectedTags) > 0) {
-                foreach ($photos as $photo) {
-                    foreach($selectedTags as $selectedTag) {
-                        if ($photoTagModel->readIsPhotoTagSetted($photo->id,$selectedTag->id)) {
-                            array_push($selectedPhotos, $photo);
-                            break;
+                $tagIds = isset ($_GET ['tags']) ? $_GET ['tags'] : "";
+
+                $tagArray = explode('-', $tagIds);
+                $tagArray = array_unique($tagArray);
+                $selectedTags = array();
+                foreach ($tagArray as $tagId) {
+                    if ($tagId != "") {
+                        if ($tagModel->readById($tagId)) {
+                            array_push($selectedTags, $tagModel->readById($tagId));
                         }
                     }
                 }
-            } else {
-                $selectedPhotos = $photos;
-            }
-
-            $photos = $selectedPhotos;
-
-            $page = 1;
-            if (isset ( $_GET ['page'] ) && $_GET['page'] * 12 - 12 <= count($photos)) {
-                $page = $_GET['page'];
-            }
-
-            $lastPic = $page * 12;
-            $firstPic = $page * 12 - 12;
-            $photosDisplay = array();
-            $amountOfPages = ceil(count($photos) / 12);
-
-            $i = 0;
-            foreach ($photos as $photo) {
-                if ($i >= $firstPic) {
-                    if ($firstPic < $lastPic) {
-                        array_push($photosDisplay, $photo);
-                        $firstPic += 1;
-                    } else {
-                        break;
+                $selectedPhotos = array();
+                if (count($selectedTags) > 0) {
+                    foreach ($photos as $photo) {
+                        foreach ($selectedTags as $selectedTag) {
+                            if ($photoTagModel->readIsPhotoTagSetted($photo->id, $selectedTag->id)) {
+                                array_push($selectedPhotos, $photo);
+                                break;
+                            }
+                        }
                     }
+                } else {
+                    $selectedPhotos = $photos;
                 }
-                $i++;
+
+                $photos = $selectedPhotos;
+
+                $page = 1;
+                if (isset ($_GET ['page']) && $_GET['page'] * 12 - 12 <= count($photos)) {
+                    $page = $_GET['page'];
+                }
+
+                $lastPic = $page * 12;
+                $firstPic = $page * 12 - 12;
+                $photosDisplay = array();
+                $amountOfPages = ceil(count($photos) / 12);
+
+                $i = 0;
+                foreach ($photos as $photo) {
+                    if ($i >= $firstPic) {
+                        if ($firstPic < $lastPic) {
+                            array_push($photosDisplay, $photo);
+                            $firstPic += 1;
+                        } else {
+                            break;
+                        }
+                    }
+                    $i++;
+                }
+                $photos = $photosDisplay;
+
+                $tags = $tagModel->readAllTags();
+
+                $view = new View('general/main_start', array("heading" => "Album"));
+                $view->display();
+                $view = new View('album/index', array("album" => $album, "tags" => $tags, "selectedTags" => $selectedTags, "tagIds" => $tagIds, "photos" => $photos, "amountOfPages" => $amountOfPages, "selectedPage" => $page));
+                $view->display();
+                $view = new View('general/main_end');
+                $view->display();
+            } else {
+                header ( 'Location: /albums' );
             }
-            $photos = $photosDisplay;
-
-            $tags = $tagModel->readAllTags();
-
-            $view = new View('general/main_start', array("heading" => "Album"));
-            $view->display();
-            $view = new View('album/index', array("album" => $album, "tags" => $tags, "selectedTags" => $selectedTags, "tagIds" => $tagIds, "photos" => $photos, "amountOfPages" => $amountOfPages, "selectedPage" => $page));
-            $view->display();
-            $view = new View('general/main_end');
-            $view->display();
         } else {
             header ( 'Location: /home' );
         }
@@ -143,14 +149,20 @@ class AlbumController extends Controller
 
         if (isset ( $_SESSION ['loggedIn'] ) && $_SESSION ['loggedIn'] == true) {
             $albumModel = new AlbumModel();
-            $album = $albumModel->readById($id);
+            $userModel = new UserModel();
 
-            $view = new View('general/main_start', array("heading" => "Album edit"));
-            $view->display();
-            $view = new View('album/edit',array("album" => $album));
-            $view->display();
-            $view = new View('general/main_end');
-            $view->display();
+            if ($albumModel->readIsAlbumFromUser($id,$userModel->readIdByUsername($_SESSION['userName']))) {
+                $album = $albumModel->readById($id);
+
+                $view = new View('general/main_start', array("heading" => "Album edit"));
+                $view->display();
+                $view = new View('album/edit', array("album" => $album));
+                $view->display();
+                $view = new View('general/main_end');
+                $view->display();
+            } else {
+                header ( 'Location: /albums' );
+            }
         } else {
             header ( 'Location: /home' );
         }
@@ -162,20 +174,24 @@ class AlbumController extends Controller
         }
 
         if (isset ( $_SESSION ['loggedIn'] ) && $_SESSION ['loggedIn'] == true) {
-            if (isset ($_POST ['editAlbum'])) {
-                $formValues = $this->getFormValues();
+            $albumModel = new AlbumModel();
+            $userModel = new UserModel();
+            if ($albumModel->readIsAlbumFromUser($id,$userModel->readIdByUsername($_SESSION['userName']))) {
+                if (isset ($_POST ['editAlbum'])) {
+                    $formValues = $this->getFormValues();
 
-                if ($this->isFieldValid("/^[a-zA-Z0-9. ]{1,45}$/", $formValues['name'])) {
-                    $albumModel = new AlbumModel();
-                    $userModel = new UserModel();
-                    $albumModel->updateNameById($formValues['name'], $id);
+                    if ($this->isFieldValid("/^[a-zA-Z0-9. ]{1,45}$/", $formValues['name'])) {
+                        $albumModel->updateNameById($formValues['name'], $id);
 
-                    header ( 'Location: /album/index/'.$id );
+                        header('Location: /album/index/' . $id);
+                    } else {
+                        header('Location: /album/edit/' . $id);
+                    }
                 } else {
-                    header ( 'Location: /album/edit/'.$id );
+                    header('Location: /album/edit/' . $id);
                 }
             } else {
-                header ( 'Location: /album/edit/'.$id );
+                header ( 'Location: /albums' );
             }
         } else {
             header ( 'Location: /home' );
@@ -189,8 +205,13 @@ class AlbumController extends Controller
     public function doDelete($id) {
         if (isset ( $_SESSION ['loggedIn'] ) && $_SESSION ['loggedIn'] == true) {
             $albumModel = new AlbumModel();
-            $albumModel->deleteById($id);
-            header('Location: /albums');
+            $userModel = new UserModel();
+            if ($albumModel->readIsAlbumFromUser($id,$userModel->readIdByUsername($_SESSION['userName']))) {
+                $albumModel->deleteById($id);
+                header('Location: /albums');
+            } else {
+                header ( 'Location: /albums' );
+            }
         } else {
             header ( 'Location: /home' );
         }
@@ -198,7 +219,7 @@ class AlbumController extends Controller
 
     private function getFormValues() {
         $values = array (
-            'name' => (isset ( $_POST ['name'] ) ? $_POST ['name'] : "")
+            'name' => (isset ( $_POST ['name'] ) ? htmlspecialchars($_POST ['name']) : "")
         );
         return $values;
     }
